proptalk/server/requirements.txt
flask==3.1.0
flask-cors==5.0.0
flask-socketio==5.4.1
openai-whisper==20240930
torch>=2.0.0
psycopg2-binary==2.9.10
PyJWT==2.10.1
google-auth==2.38.0
google-api-python-client==2.159.0
gunicorn==23.0.0
eventlet==0.37.0
requests==2.32.3
anthropic>=0.40.0
APScheduler==3.10.4
cryptography>=42.0.0
